Apply a function to each row of a data.table and add the result as a column

Apply a function to each row of a data.table and add the result as a column

benchmark(f, benchdata, fname = NULL, table.file = NULL,
  report = function(p, obj) { }, result.only = FALSE,
  id.result.only = FALSE, paramsEvalAtCall = "", ...)

Arguments

f
a character or a function object

benchdata
a data.table, can be NULL in which case the table.file is loaded expecting it to contain a data.table object.

fname
a character indicating the column-name to be used for the column of result, can be NULL in which case the name of the function specified by f is used.

table.file
a character specifying a .RData file where the resulting data.table object is to be stored. Defaults to NULL. See benchdata.

report
a reporting function (p, obj) called after executing f on a row, used for logging purpose.

result.only
logical indicating whether only the list of result objects should be returned, default FALSE

id.result.only
logical indicating whether the returned data.table shall be limited to the id and fname columns only, default FALSE. Ignored if result.only is specified as TRUE.

paramsEvalAtCall
a character, either an empty string '', or an R expression or comma separated expressions such as 'parFixed=c(theta=mean(p$v))), parMax=c(theta=max(p$v))'. At evaluation these expression have access to the list p that is passed to f.

...
additional parameters passed to f

Value

If result.only is FALSE (default) a data.table object equal to the original benchdata with a column named as fname for the results of calling f. If a column of the same name already exists, it is overwritten. If the parameter id.result.only is FALSE (default) all the remaining columns are also contained in the returned data.table. Otherwise only the columns id, and fname are returned. If result.only is TRUE, than only a list of the resulting objects for each row of the original data.table is returned.