Parallelize a benchmark on data.table by calling benchmark on equal subsets of its lines.

Parallelize a benchmark on data.table by calling benchmark on equal subsets of its lines.

genBenchJobs(f, fname = NULL, script.file = fname, table.file,
  table.name = "NULL", ids = NULL, params = "",
  paramsEvalAtCall = "", perJob = 10, type = c("shell", "bsub",
  "qsub"), command = NULL, requires = c(), sources = c(),
  code = "", bsub.W = "12:00", bsub.n = 1, bsub.mem = 1000,
  bsub.other = "", qsub.q = "sc02.q", sleep.every = 50,
  sleep.secs = 300)

Arguments

f
a character or a function object called on each line of benchdata.

fname
a character indicating the name of the column to be filled.

script.file
a character containing a part of the file-name of the generated scripts:

table.file
a character indicating the name of a .RData file containing a data.table object.

ids
a vector indicating a possible subset of the id-column of benchdata on which to perform the benchmark. Defaults to NULL in which case all lines are consiedered.

params
a character containing the additional parameters to f as they should be written in the call to it, ex. "foo1=5, foo2=6".

paramsEvalAtCall
a character, either an empty string '', or an R expression or comma separated expressions such as 'parFixed=c(theta=mean(p$v))), parMax=c(theta=max(p$v))'. At evaluation these expressions have access to the list called p which is passed as a first argument to f. This character string should not contain double-quotes.

perJob
number of lines per subset, defaults to 10.

type
the type of command to generate, defaults to 'shell'. Currently, should be one of 'shell', 'bsub', 'qsub'.

command
a character indicating a custom command to be used, defaults to NULL. If specified, the type parameter is ignored.

requires
a vector of strings indicating R-pacages to be loaded prior to executing the benchmark

sources
a vector of strings indicating R-files to be sourced prior to executing the benchmark.

code
a string containing any r-code that is to be prepended to the R-script before the call to f.

bsub.W
a character indicating job-waiting time if type 'bsub', defaults to '12:00'

bsub.n
a character indicating job-number of processes if type ='bsub', defaults to 1.

bsub.mem
an integer indicating number of megabites used by a job if type = 'bsub', defaults to 1000.

bsub.other
Character containing other options of bsub to be written first.

qsub.q
a character indicating queuename if type='qsub', defaults to 'sc02.q'

sleep.every
integer, number of job-submit commands between a sleep command

sleep.secs
integer, number of seconds to sleep after every sleep.every job-submissions

Value

a character vector containing the generated shell commands (see details).

Details

This function generates two files: 1. an R-script, j_<script.file>_.R, which exctracts table.file and line-numbers from the command-line arguments of the R-process and calls benchmark executing f on those lines; 2. a shell script, j_<type>_<script.file>.sh of commands for starting an R process and executing the R-script on specific lines of the data.table.